refactor: collapse manifest-formatter parent to wiring facade (resolves #1871)
